test_schema_normalize_expr_with_equivalence
